rename the trigger item components
