+++ openmetadata-ui/src/main/resources/ui/src/pages/AddObservabilityPage/AlertFormSourceItem/AlertFormSourceItem.interface.ts
@@ -13,6 +13,6 @@
 
 import { FilterResourceDescriptor } from '../../../generated/events/filterResourceDescriptor';
 
-export interface ObservabilityFormTriggerItemProps {
+export interface AlertFormSourceItemProps {
   filterResources: FilterResourceDescriptor[];
 }
